move modern TransactionStatusIcon & TransactionStatusText
