+++ src/components/HistoryEventDetails/layouts/TokenExchangeEventDetails.js
@@ -31,8 +31,8 @@ import { Row, ColumnRight, Spacing } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
 import TokenValueView from 'components/display/TokenValueView';
-import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
-import TransactionStatusText from 'components/modern/TransactionStatusText';
+import TransactionStatusIcon from 'components/display/TransactionStatusIcon';
+import TransactionStatusText from 'components/display/TransactionStatusText';
 
 // Constants
 import { EXCHANGE } from 'constants/navigationConstants';
