+++ src/components/HistoryEventDetails/layouts/TokenTransactionEventDetails.js
@@ -31,8 +31,8 @@ import { Row, Spacing } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
 import TokenValueView from 'components/display/TokenValueView';
-import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
-import TransactionStatusText from 'components/modern/TransactionStatusText';
+import TransactionStatusIcon from 'components/display/TransactionStatusIcon';
+import TransactionStatusText from 'components/display/TransactionStatusText';
 
 // Constants
 import { SEND_TOKEN_FROM_HOME_FLOW } from 'constants/navigationConstants';
