+++ src/components/HistoryEventDetails/layouts/WalletEventDetails.js
@@ -30,8 +30,8 @@ import Modal from 'components/Modal';
 import ReceiveModal from 'screens/Asset/ReceiveModal';
 import Text from 'components/core/Text';
 import Toast from 'components/Toast';
-import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
-import TransactionStatusText from 'components/modern/TransactionStatusText';
+import TransactionStatusIcon from 'components/display/TransactionStatusIcon';
+import TransactionStatusText from 'components/display/TransactionStatusText';
 
 // Actions
 import { viewTransactionOnBlockchainAction } from 'actions/historyActions';
