+++ src/components/HistoryList/items/HistoryListItem.js
@@ -26,7 +26,7 @@ import { SvgCssUri } from 'react-native-svg';
 import Icon, { type IconName } from 'components/core/Icon';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
-import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
+import TransactionStatusIcon from 'components/display/TransactionStatusIcon';
 
 // Utils
 import { useThemeColors } from 'utils/themes';
