+++ src/components/display/TransactionStatusIcon.js
